update getUserByGuid fn

+++ routes/v1/users/users.js
@@ -538,7 +538,10 @@ router.route("/delete")
 
     params.validate()
       .then(() => {
-        // just to be sure that user exist in DB
+        // there might be cases when user exists in Auth0 but doesn't exist in MySQL
+        if (transformedParams.guid === '') {
+          return true;
+        }
         return usersService.getUserByGuid(transformedParams.guid, true);
       })
       .then(() => {
@@ -548,6 +551,9 @@ router.route("/delete")
         return auth0Service.deleteAuth0User(token, transformedParams.user_id);
       })
       .then(() => {
+        if (transformedParams.guid === '') {
+          return true;
+        }
         return usersService.removeUserByGuidFromMySQL({ guid: transformedParams.guid });
       })
       .then(() => {
